fix: move tests to atexit
